+++ new.js
@@ -1,0 +1,2 @@
+const { MongoClient } = require("mongodb");
+MongoClient.connect("")
